# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: SIGP (Sistema Integral de Gestión de Proyectos) API
  version: 1.0.0
  description: |
    API REST completa para el Sistema Integral de Gestión de Proyectos (SIGP).

    El sistema gestiona:
    - Autenticación y autorización basada en roles
    - Planificación estratégica (PGD, OEI, OGD, OEGD, Acciones Estratégicas)
    - Proyectos y subproyectos (POI)
    - Gestión ágil (Épicas, Sprints, Historias de Usuario, Tareas)
    - Recursos humanos (Personal, Divisiones, Habilidades)
    - Notificaciones en tiempo real
    - Gestión de archivos con MinIO
    - Dashboards y métricas

  contact:
    name: API Support
    email: api@inei.gob.pe

  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:3010/api/v1
    description: Development server
  - url: https://api.sigp.inei.gob.pe/api/v1
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtenido en el login
  schemas:
...
    LoginDto:
      type: object
      required:
        - password
      properties:
        email:
          type: string
          format: email
          example: usuario@inei.gob.pe
          description: Email del usuario (requerido si no se proporciona username)
        username:
          type: string
          minLength: 3
          maxLength: 100
          pattern: '^[a-zA-Z0-9_-]+$'
          example: jperez
          description: Nombre de usuario (requerido si no se proporciona email)
        password:
          type: string
          minLength: 6
          example: password123
          description: Contraseña del usuario
      oneOf:
        - required: [email, password]
        - required: [username, password]
    RegisterDto:
      type: object
      required:
        - email
        - username
        - password
        - nombre
        - apellido
        - rol
      properties:
        email:
          type: string
          format: email
          example: usuario@inei.gob.pe
          description: Email del usuario
        username:
          type: string
          minLength: 3
          maxLength: 100
          pattern: '^[a-zA-Z0-9_-]+$'
          example: jperez
          description: Nombre de usuario único (solo letras, números, guiones y guiones bajos)
        password:
          type: string
          minLength: 6
          example: password123
          description: Contraseña del usuario
        nombre:
          type: string
          example: Juan
          description: Nombre del usuario
        apellido:
          type: string
          example: Perez
          description: Apellido del usuario
        rol:
          type: string
          enum:
            - ADMIN
            - PMO
            - COORDINADOR
            - SCRUM_MASTER
            - PATROCINADOR
            - DESARROLLADOR
            - IMPLEMENTADOR
          example: DESARROLLADOR
          description: Rol del usuario en el sistema
        telefono:
          type: string
          nullable: true
          description: Teléfono de contacto (opcional)
    UserResponseDto:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          example: usuario@inei.gob.pe
        username:
          type: string
          example: jperez
          description: Nombre de usuario único
        nombre:
          type: string
          example: Juan
        apellido:
          type: string
          example: Perez
        rol:
          type: string
          enum:
            - ADMIN
            - PMO
            - COORDINADOR
            - SCRUM_MASTER
            - PATROCINADOR
            - DESARROLLADOR
            - IMPLEMENTADOR
        avatarUrl:
          type: string
          nullable: true
    AuthResponseDto:
      type: object
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        expiresIn:
          type: integer
          description: Segundos hasta que expira el token de acceso
          example: 3600
        user:
          $ref: '#/components/schemas/UserResponseDto'
    RefreshTokenDto:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    ChangePasswordDto:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6

    # ==================== PROJECT SCHEMAS ====================
...
    CreateSprintDto:
      type: object
      required:
        - proyectoId
        - nombre
        - fechaInicio
        - fechaFin
      properties:
        proyectoId:
          type: integer
          example: 1
        nombre:
          type: string
          maxLength: 100
          example: Sprint 1
        sprintGoal:
          type: string
          nullable: true
          example: Completar módulo de autenticación
        fechaInicio:
          type: string
          format: date-time
          example: "2024-01-15T00:00:00Z"
        fechaFin:
          type: string
          format: date-time
          example: "2024-01-29T00:00:00Z"
        capacidadEquipo:
          type: integer
          minimum: 0
          nullable: true
    SprintResponseDto:
      type: object
      properties:
        id:
          type: integer
        proyectoId:
          type: integer
        nombre:
          type: string
        estado:
          type: string
          enum:
            - Planificado
            - Activo
            - Completado
        sprintGoal:
          type: string
        fechaInicio:
          type: string
          format: date-time
        fechaFin:
          type: string
          format: date-time
        capacidadEquipo:
          type: integer
        createdAt:
          type: string
          format: date-time

    # ==================== HISTORIA USUARIO SCHEMAS ====================
...
    Error:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Registrar nuevo usuario
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        201:
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
        409:
          description: Email ya registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login con email y password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        200:
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
        401:
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refrescar token de acceso
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
      responses:
        200:
          description: Token refrescado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
        401:
          description: Token de refresco inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/profile:
    get:
      tags:
        - Authentication
      summary: Obtener perfil del usuario actual
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        200:
          description: Perfil obtenido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        401:
          description: No autenticado
  /auth/change-password:
    put:
      tags:
        - Authentication
      summary: Cambiar contraseña del usuario
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordDto'
      responses:
        200:
          description: Contraseña cambiada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        400:
          description: Contraseña actual incorrecta
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Cerrar sesión del usuario
      operationId: logout
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        200:
          description: Sesión cerrada

  # ==================== PROYECTO ENDPOINTS ====================
...
  /sprints:
    get:
      tags:
        - Agile
      summary: Listar todos los sprints
      operationId: listSprints
      security:
        - bearerAuth: []
      parameters:
        - name: proyectoId
          in: query
          schema:
            type: integer
        - name: estado
          in: query
          schema:
            type: string
            enum:
              - Planificado
              - Activo
              - Completado
        - name: activo
          in: query
          schema:
            type: string
            enum:
              - "true"
              - "false"
      responses:
        200:
          description: Lista de sprints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SprintResponseDto'

    post:
      tags:
        - Agile
      summary: Crear nuevo sprint
      operationId: createSprint
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSprintDto'
      responses:
        201:
          description: Sprint creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SprintResponseDto'
  /sprints/{id}:
    get:
      tags:
        - Agile
      summary: Obtener detalles de un sprint
      operationId: getSprint
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Detalles del sprint
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SprintResponseDto'

    patch:
      tags:
        - Agile
      summary: Actualizar sprint
      operationId: updateSprint
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSprintDto'
      responses:
        200:
          description: Sprint actualizado

    delete:
      tags:
        - Agile
      summary: Eliminar sprint
      operationId: deleteSprint
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Sprint eliminado
  /sprints/{id}/iniciar:
    patch:
      tags:
        - Agile
      summary: Iniciar un sprint
      operationId: iniciarSprint
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Sprint iniciado
  /sprints/{id}/cerrar:
    patch:
      tags:
        - Agile
      summary: Cerrar un sprint
      operationId: cerrarSprint
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                razonCierre:
                  type: string
      responses:
        200:
          description: Sprint cerrado
  /sprints/{id}/burndown:
    get:
      tags:
        - Agile
      summary: Obtener gráfico burndown del sprint
      operationId: getBurndownSprint
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Datos de burndown
          content:
            application/json:
              schema:
                type: object
                properties:
                  dias:
                    type: array
                    items:
                      type: string
                  storyPointsRestantes:
                    type: array
                    items:
                      type: number
  /sprints/{id}/metricas:
    get:
      tags:
        - Agile
      summary: Obtener métricas del sprint
      operationId: getMetricasSprint
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Métricas del sprint
          content:
            application/json:
              schema:
                type: object
                properties:
                  velocidad:
                    type: number
                  storyPointsPlaneados:
                    type: integer
                  storyPointsCompletados:
                    type: integer
                  tasasCompletitud:
                    type: number

  # ==================== HISTORIA USUARIO ENDPOINTS ====================
...
  /dashboard/proyecto/{id}/velocidad:
    get:
      tags:
        - Dashboard
      summary: Velocidad del equipo del proyecto
      operationId: getVelocidadProyecto
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Datos de velocidad